refactor (notification): move creating listener for GetAllNotifications to repository
