+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/domain/repositories/NotificationRepository.kt
@@ -1,5 +1,7 @@
 package com.personalization.sdk.domain.repositories
 
+import com.personalization.api.OnApiCallbackListener
+import com.personalization.api.responses.notifications.GetAllNotificationsResponse
 import com.personalization.sdk.domain.models.NotificationSource
 import org.json.JSONObject
 
@@ -20,4 +22,9 @@ interface NotificationRepository {
         page: Int?,
         limit: Int?
     ): JSONObject
+
+    fun getAllNotificationListener(
+        onGetAllNotifications: (GetAllNotificationsResponse) -> Unit,
+        onError: (Int, String?) -> Unit = { _: Int, _: String? -> }
+    ): OnApiCallbackListener
 }
